refactor : domain은 domain 패키지 안으로 이동
